+++ lib/load.php
@@ -26,17 +26,24 @@ class Load {
 
 	function findByRegistry($template) {
 		$blueprints = kirby()->get('blueprint');
-		if(isset($blueprints[$template]) && f::exists($blueprints[$template])) {
-			return array($template, $blueprints[$template]);
+		if(isset($blueprints[$template])) {
+			if(f::exists($blueprints[$template])) {
+				return array($template, $blueprints[$template]);
+			} elseif(f::exists($blueprints['default'])) {
+				return array($template, $blueprints['default']);
+			}
 		}
 	}
 
 	function findByRoot($template) {
 		$filepath = $this->kirby->roots->blueprints . DS . $template . '.yml';
+		$defaultpath = $this->kirby->roots->blueprints . DS . 'default.yml';
 		$array = array($template => $filepath);
 
 		if(f::exists($filepath)) {
 			return array($template, $filepath);
+		} elseif(f::exists($defaultpath)) {
+			return array($template, $defaultpath);
 		}
 	}
 }