+++ docs/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+**0.3**
+
+- Fixed major bug with singleton class.
+- Fixed major bug with `bread::parse()`.
+- Added support for global field definitions in structure fields.
+- Added support for global field definitions extends in structure fields.
+- Added [compare](compare.md) to Kirby Architect.
+
 **0.2**
 
 - Complete rewrite